Version final

+++ core/src/com/diamon/personaje/JefeDos.java
@@ -103,7 +103,7 @@ public class JefeDos extends Personaje {
 
 			y = pocicionY + distanciaMovimientoY + (distanciaMovimientoY * MathUtils.sinDeg(tiemo * velocidadY));
 
-			x += velocidadX / Juego.DELTA_A_PIXEL * delta;
+			x += (int) velocidadX / Juego.DELTA_A_PIXEL * delta;
 
 			if (x >= camara.position.x + (Juego.ANCHO_PANTALLA / 2 - getWidth())
 					|| x <= camara.position.x - (Juego.ANCHO_PANTALLA / 2 - 200)) {
