+++ core/src/com/diamon/personaje/JefeUno.java
@@ -78,7 +78,7 @@ public class JefeUno extends Personaje {
 
 			y = pocicionY + distanciaMovimientoY + (distanciaMovimientoY * MathUtils.sinDeg(tiemo * velocidadY));
 
-			x += velocidadX / Juego.DELTA_A_PIXEL * delta;
+			x += (int) velocidadX / Juego.DELTA_A_PIXEL * delta;
 
 			if (x >= camara.position.x + (Juego.ANCHO_PANTALLA / 2 - getWidth())
 					|| x <= camara.position.x - (Juego.ANCHO_PANTALLA / 2 - 200)) {
